--- .bundle/config
@@ -1,2 +1,0 @@
----
-BUNDLE_WITHOUT: production
