+++ bin/config_example.yml
@@ -30,9 +30,11 @@ buns_list:
 
 ReplayTable:
   self:
-    - 
-      - [!ruby/regexp /(oe|おーいー)(_||\s)(bot|ボット|ﾎﾞｯﾄ|ぼっと)/i]
-      - ["はい。"]
+    - !ruby/regexp /(oe|おーいー)(_||\s)(bot|ボット|ﾎﾞｯﾄ|ぼっと)/i
+    - ["はい。","呼びましたか。","呼ばれた気がしました。"]
+  call:
+    - !ruby/regexp /^(oe|おーいー)(_||\s)(bot|ボット|ﾎﾞｯﾄ|ぼっと)$/i
+    - ["はい。","呼びましたか。","呼ばれた気がしました。"]
 
   comprehensible:
     - 
@@ -51,10 +53,10 @@ ReplayTable:
       - ["こんにちは"]
       - ["こんにちは。","こんにちはです。"]
     - 
-      - ["おはよ","おきた","起きた"]
+      - ["おはよ",!ruby/regexp /(お|起)きた/]
       - ["おはようございます。","おはようです。"]
     - 
-      - ["おやすみ","寝る","ねる","寝ます","ねます"]
+      - ["おやすみ",!ruby/regexp /(寝|ね)(る|ます)/,"ねむい","眠い"]
       - ["おやすみなさい。","また明日です。","おやすみです。"]
     - 
       - ["本当","ホント","ほんとう","うそ","嘘"]
@@ -65,6 +67,12 @@ ReplayTable:
     - 
       - ["おっぱい"]
       - ["そんなにおっぱいが好きなんですか。","\n( ﾟ∀ﾟ)o彡°おっぱい！おっぱい！"]
+    - 
+      - ["えらい","偉い","てんさい","天才","すごい","あたまいい","頭いい"]
+      - ["えへへ。","もっとほめてください。","照れます。"]
+    - 
+      - ["つらい","辛い","もうダメだ","もうだめだ",!ruby/regexp /(死|し)(ぬ|にたい|のう)/]
+      - ["\nがんばれ♡がんばれ♡です。","つよく生きてください。","つらそうです。"]
     - 
       - ["わかる"]
       - ["わかられました。"]
@@ -96,6 +104,18 @@ ReplayTable:
       - ["線形代数","せんけいだいすう"]
       - ["あなたの線形代数の単位はDです。"]
     - 
+      - ["死ね","しね"]
+      - ["生きます。"]
+    - 
+      - ["生きてる","死んでる"]
+      - ["生きてます。"]
+    - 
+      - ["おい","あの","もしもし"]
+      - ["なんですか。","何かごようですか。"]
+    - 
+      - [!ruby/regexp /ping/i]
+      - ["pong"]
+    - 
       - ["機能"]
       - ["https://github.com/TKB-AEM/OE_bot/blob/master/README.md"]
     - 
@@ -110,4 +130,4 @@ ReplayTable:
     - "\n（ ˘⊖˘）。o(何言ってるんだこの人)"
     - "\nわかります。（わからない顔）"
     - "日本語でおｋです。"
-    - "わけがわからないよです。"
+    - "わけがわからないです。"
