+++ lib/oebot/database/helper.rb
