+++ lib/oebot/function/helper.rb
